For OSS signing move sign.targets from BeforeCommon..to BeforeCSharp..
For OSS signing, product and test assemblies need to have a defined
constant so they know if they are being signed.  They use this to know
if they should refer to each other by simple or by strong name.  This
change causes sign.targets to be included in the product/test assembly
build in CustomBeforeMicrosoftCSharpTargets instead of
CustomBeforeMicrosoftCommonTargets, which causes the injected changes
to the DefineConstants MSBuild property to be picked up during build.