Make the build tools repo build consistently.
This change mirrors the changes maded recently in the corefx repo to make it
use dir.props and dir.targets in each of the projects. To ensure a consistent build.

This cleans up a few targets files that we no longer need after we switched
to the dir.props/dir.targets/dirs.proj model (like fxbuild.*, and beforecommon).

In order to have a consistent build one additional change needed was to put the
version for the BuildTools package in the nuspec as we don't want to load the
task library we are packaging to try and find a package version.