Port "xunit.console" project to new "xunit.console.netcore" project
This project is designed to target .NET Core.

It depends on other runtime components of xunit including
xunit.execution and xunit.runner.utility.  Those are similarly being
modified, but will be done in a fork of the main xunit repository, so
that the changes can be merged upstream.  Those changes can be merged
upstream because they depend on contracts that are available in
portable targeting packs, and thus can be built by targeting an
appropriate portable profile.

This xunit.console.netcore project is being added to the buildtools
repository because it depends on a contract (System.Console) which is
not a part of any currently available portable targeting pack.