Clean up depending.targets file to make it more general purpose
Renamed the target to make it more descriptive to RestoreAllPackageDependencies
while keeping the old EnsureDependencies target as a wrapper for now.

Added inputs/outputs to make incremental build work without requiring the
restore of all the packages on every build.

Added an empty PackageConfigsToExclude so people can override the wildcard Include
if needed.

Started outputing a message whenever the target starts as it tends to take a little
while and it is useful for people to know that restore is happening.