Create the infrastructure to allow testing changes across repos.
This is going to be driven by the localpublish.props file at the source root
of the repo. By default this file does not exist and it is not tracked by git.

You can create the file with contents similar to this to enable the repo that
is currently built from publishing to a local path. This enables the scenario
when one package has a dependency to the other and you want to test your
changes locally.

If the file is present the following are going to happen:
- The package folder in the target repo is going to be removed
- A packages.config file is going to be created for the package that we are
installing. This is required to bypass the resolving of dependencies that is
triggered by installing a packaged referenced by name.
- A nuget install command is issued to install the package in the target repo

Structure of localpublish.props file:
<Project ToolsVersion="12.0" DefaultTargets="Build"
         xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
  <ItemGroup>
    <LocalPackages Include="<directory of the package created by the repo>">
      <PackageName><name of the package></PackageName>
      <PackageVersion><version, including -prerelease></PackageVersion>
      <InstallLocation><Location in the other repo where the package
      					should be installed</InstallLocation>
    </LocalPackages>
</Project>