Create Build.Common.targets that is the one import to rule them all.
This change adds a Build.Common.targets that will be imported by
individual repo dir.targets files to get all the individual project
targets needed. It pulls back in a lot of the changes that have
been made directly in corefx repo dir.targets and makes them more
generalized for other repo's to consume the same functionality.

The basic idea is to try and minimize the boot strapping code and
eliminate duplication for all the repos that will be using
the build tools package.