Fix issues with BuildNumber property.
The VSO official build defines global property BuildNumber which is
colliding with our property with the same name.  I have renamed the
property to RevisionNumber to avoid the collision.

Rename BuildNumber.props to the more generic BuildValues.props as it's
likely we'll have other build values in the in the future.

Rename UpdateBuildNumber.targets to UpdateBuildValues.targets.

Sequence GetNextRevisionNumber before target BuildAllProjects for corefx
builds.