Fix XunitTraitOptions to be a replacement instead of appending them
When code coverage is enabled we cannot just append the command
line options to the end of the command line because the have to be
explicitly quoted and passed to the code coverage host. Instead we
put a placeholder into the command line that we replace later with
the real trait options. This allows us to keep the code coverage and
the testing targets decoupled and still be able to compute the
traits as part of the target.