Add initial set of perf tools.
This adds project and NuGet package infrastructure for bring up perf
tooling.  At present the tooling is split into a platform-specific
collection mechanism and a platform-agnostic model of the data used for
reporting purposes.

Created a new NuGet package Microsoft.DotNet.PerfTools to contain all of
the perf tools.

EventTracer
This is a console app used to start and stop collection of ETW data and
can optionally generate an XML report of the data collected.  Since it
relies on ETW it is applicable to Windows only.

PerfEventsData
This library contains a platform agnostic model of performance data.  At
present it only contains performance data types applicable to our
"consumptive" metrics.

ComparePerfEventsData
This is a console app for comparing baseline and live performance data.

GenPerfBaseline
Console tool that will merge multiple result sets to generate a baseline.