Perf tools fixes.
Increase event buffer size to 512MB (the default is 64MB).

Write a warning message with any lost events; this happens when the event
buffer is too small.

Throw an exception is report generation failed to parse any data for the
specified process name.

Aggregate data for GC object allocation events.

Add text for exceptions thrown in GenPerfBaseline to clarify failures.