Fix missing coverage dependency
If the coverage.xml got deleted after running coverage once, and then the build was run with coverage again, it would not rerun the actual coverage step.  It still would attempt to recreate then coverage report, which would fail because of a missing coverage file.  Fix this by adding the coverage output to the outputs of the test target (only set when coverage is set to true).