Workaround issue where no reference is selected for a library with only lib.
Currently our msbuild task cannot distinguish between lib only package and lib with a placeholder in ref.  The former should use the lib as a reference assembly but does not.   Temporarily add it for packages that we know are not using placeholders in ref.  We should remove this workaround once we can distinguish between no-ref and ref with placeholder.