Prevent concurrent restore of same project.json
DNX has an issue with concurrent restore: https://github.com/aspnet/dnx/issues/1682

Previously I would workaround by copying to intermediate, but this breaks in official build where intermediate is not under the scope of our nuget.config, so custom package sources are ignored.

Fix this instead by wrapping the call to dnu restore with a mutex unique to the project.json name.  We only need to do this for the test project.json because that is the only one we'll restore concurrently.