Integrate latest drop of ResolveNuGetPackageAssets task
This change eliminates the majority of the code from ResolveNuGetPackageAssets and relies on the applicability calculated during restore and saved to the lock file.
Restore now supports all of the folder formats we depend on for V2 PCLs so I can eliminate packages.config.  For now I have only eliminated the test packages.config, and will leave support in for resolving from packages.config until all upstream repos are off of packages.config.