Always run the package restore step
This is specifically targeted toward the new json package format.  The issue we have is that the coreclr tests have project names that sometimes non-unique in the tests (there are template projects).  This causes locking issues when restoring packages, or occasionally packages do not get restored at all.  This is because the semaphore file is based off of the project name.  This implicitly adds a restriction that project file names must be unique within an msbuild invocation.  The semaphore file was an attempt to have incremental package restore.  Under  dnx, the lock files dramatically speed up restore and so always restoring is okay.

This changes removes the incremental restore and unblocks coreclr's move to dnx.