Add support for compiling CIL files.
This change introduces a targets file to support projects that are
written in MSIL. Like Microsoft.{CSharp,VisualBasic}.targets, the
targets file builds on the infrastructure provided by
Microsoft.Common.targets, which provides the usual Build, Clean,
etc. targets. Currently the only tool that this file understands
is ILAsm, so compilation will fail on platforms where this tool is
not available.