Move test deployment to build phase and speed it up
The experience around building and running tests in VS got broken by a
change to only copy tests when they're run from msbuild. Move them back
to PrepareForRun to bring back the VS workflow.

In doing so, it became apparent that the test deployment step was
copying files on every invocation regardless of whether there were any
changes. That turned out to be because there are different versions of
files with the same simple name destind for the same output folder. As
such, every iteration would swap them back and forth. The fix here is
too trim out the duplicates before invoking the Copy task. Some of the
duplication is by design (e.g. to override nuget assets with local
assets), but we are also getting multiple nuget package versions out
of the nuget asset resolution, which seems like a bug, which is now filed
as #182.

To further speed things up, using hard-links is now always on for the
test deployment step, which currently copies tons of the same files
to each test folder.

Fix #181