Add a build for GenFacades.exe and dependencies
This adds a GenFacades build for both Desktop (.NETFramework 4.6) and
CoreCLR (using CoreConsole.exe as a host). There is an extra targets file
(publishexe.targets) which handles some of the extra configuration needed
to create a CoreCLR-based Console application, including the runtime,
host, framework libraries, etc. I've included nuspec files to create
nupkg's for both versions of GenFacades (CoreCLR and Desktop) which contain
fully-standalone tools. The Desktop verison currently requires .NET 4.6
since it compiles against the contracts and therefore needs some of the
in-box facades.