Fix tests running under code coverage
Under code coverage, a System.Diagnostics.Process test was failing because the invoked process name did not match in casing to the expected value.  The expected value matches with the casing of CoreRun.exe, while the test was seeing corerun.exe on Windows.  The issue here appears to be that under OpenCover, the invoked target process (corerun) doesn't have its name canonicalized vs. the file system.  Fix this by setting the test host for Windows and Linux with the correct casing.