Show _AssemblyInfo.cs in Solution Explorer
The Multilingual App Toolkit (MAT) requires the AssemblyInfo file to be
included in EnvDTE so that it can read the assembly's neutral language.
This change sets Visible to true for the AssemblyInfo file and gives it
an appropriate Link.  It also moves it outside of the
GenerateAssemblyInfo target, because it only is shown in Solution
Explorer if it's not in a target.