Generate _AssemblyInfo.cs in neutral output path
_AssemblyInfo.cs was previously generated in a configuration-specific
path, such as bin\obj\FreeBSD.AnyCPU.Debug\System.Console.  For projects
which build with an architecture other than Windows_NT.AnyCPU.Debug, the
_AssemblyInfo.cs link in Solution Explorer was broken because it was
pointing to Windows_NT.AnyCPU.Debug.

Because _AssemblyInfo.cs is not specific to the OS or configuraiton,
this change fixes the link by moving _AssemblyInfo.cs to a neutral
output path.