Fix RAR warnings when building ReferenceAssemblies.
Implementation dependencies of portable contracts were leaking into the reference assembly
build resulting in warnings from RAR.  This happens because we are resolving packages against
'dotnet' and a package with a dotnet implementation will also have its implementation
dependencies which may be larger than the reference assembly dependencies.

Add a feature to the MSBuild task to support only returning direct dependencies for compile
assets and enable it for reference assemblies.