Change the path convention for reference assemblies
Previously the convention was ref\major.minor\assembly.csproj.  This will lead to poor history
over time since we'll have to source every time we make an API change to a stable package.

Instead, default to ref\assembly.csproj, and only create ref\major.minor\assembly.csproj if
we need to build an older version simultaneous with the latest package.