Fix rebuild of partial facade with no ProjectReferences
When running a rebuild of a partial facade project Clean would evaluate the condition
on AssignProjectConfiguration before running our target.  This resulted in skipping
AssignProjectConfiguration.  Then _SplitProjectReferencesByFileExistence would run
and operate on an empty set of project references (since AssignProjectConfiguration
was skipped).  Later on during build, our target would run, but
_SplitProjectReferencesByFileExistence would not since it already ran.  The net result
was that even though we added the project reference it wasn't being consumed.

Fix this by changing the scheduling of FindPartialFacadeProjectReference to ensure
it runs in both build and clean.