Add common localization code
This localization code is common to all C# projects, and is the code
added by the Multilingual App Toolkit (MAT) for maintaining localization
in a project.

When building a localized project on a machine with MAT installed, MAT
synchronizes the .xlf files in the project with the Strings.resx, and
synchronizes the localized Strings.resx files with the .xlf files.  If
MAT is not installed on the machine, the loc file synchronization is
skipped and a message is written to the log file.

If a project does not have a .de.xlf file, then by default the localization target is not imported.  This behavior can be overriden using the ExcludeLocalizationImport property.

For the files to show up in Solution Explorer, it is necessary to set
Visible to True.