Remove failing Error in GenerateAssemblyInfo
With https://github.com/dotnet/buildtools/commit/a85967b2aab1bbed12c2dcb523cba9f0d4e51c8b, the Error
in GenerateAssemblyInfo was changed to occur when AssemblyInfoFileDir
does not exist.  However, in the case where AssemblyInfoFileDir is set
to "$(BaseIntermediateOutputPath)$(MSBuildProjectName)\", it is expected
to not exist, and is created by a MakeDir task later in the target.

With the fix to https://github.com/dotnet/buildtools/issues/231, this file will probably be reverted
back to how it was before the _AssemblyInfo.cs changes made for loc,
meaning that the Error will look for the existence of
$(IntermediateOutputPath).

Until that fix, though, I think the right workaround is to remove the
Error.