Disable "warning" for no tests in assembly
The xunit.console.netcore runner is printing out "Warning: " when there are no tests to run in the specified assembly.  This causes msbuild to treat that as a warning, which gets flagged and accumulated in the warnings statistics when doing a full build.  That results in 17 warnings when building corefx today, which is very noisy and ends up making it difficult to spot real warnings that should be fixed.

This commit simply changes the string to be "Info" instead of "Warning", and also uses a slightly different shade of yellow in order to differentiate it from more important warnings.  This brings the number of warnings generated during a full corefx build back down to 0.