Create a nuget package for GenAPI
Referencing this package brings in a targets file that emits a CS with file to the output
directory (overridable to any path chosen by the project).  In practice I expect projects
to override this to point to a location under source control in another project that builds
the reference assembly.

This can be used to automate the production/update of reference assemblies.
I've chosen to emit source since any API diffs will clearly show up in source control.