Support partial facade generation on X-Plat
This change does the minimial amount of work to get partial facade
generation into a state where it can be run as part of the build when
running on Linux or OSX.

The major problem was that since we don't produce PDBs when building
with Roslyn on Linux or OSX, GenFacades needs to be able to handle the
case where PDBs are missing.  The version of GenFacades that we had in
build tools can handle this (it issues a warning when doing so) but we
actually don't consume this version from BuildTools today. This change
addresses that issue.

 - Update partialfacades.targets to consume the GenFacades.Desktop
   pacakge.  In doing so, I have introduced a new targets file that
   ships as part of build tools that has version number information
   about the revision of build tools in question so we don't have to
   hardcode the ever changing revision into the targets.  This does
   require that folks that restore the build tools and want to use
   partial facades need to now start restoring the GenFacades.Desktop
   pacakge as well.

 - Fix up some logic in the targets to not try to move PDBs if
   DebugSymbols is false, as it is on Linux and OSX.

 - Add a way to supress the warning to GenFacades for the case where the
   PDB is known not to exist

With these changes, GenFacades can run and fill partial facades on Linux
when running on a custom version of Mono (Mono is missing
Array.Empty<T>() and some StringBuilder overloads which
Microsoft.Cci.dll depends on, these will be added when Mono consumes an
updated reference assembly drop, for now it is easy enough to add them
and then build locally if you really want this to work, which I do.
Longer term we should just move to running GenFacades.exe on top of
CoreCLR, once we have cross platform coreclr runtime packages we can
consume.)