Remove tool runtime hardlinks and add extensibility for cmd
We believe hardlinks make for an interesting race condition in some cases
while writing to the same location. Given that this is only copied to
the one location we will disable the hardlinks to see if it helps with
reliability.

While running CoreRun in a build enviroment that as COMPLUS_InstallRoot
set the COM activation for things like the PDB APIs fail to activate
because the runtimes don't match. To help eliminate this problem
we clear the COMPLUS_InstallRoot before calling CoreRun. This new
ToolHostCmd also allows folks to also add other such env variable tweaks
as part of the command if necessary.