Fix code coverage report generation
Code coverage is failing because the path to the report generator tool is being created incorrectly.  It appears this is due to the version number of the report generator changing from 2.3.0.0 to 2.3.0.  As long as I was fixing that, I updated it to the most recent version, 2.3.1.  I verified code coverage works now with the change.