Ensure we don't copy runtime assets for class libs
My change to use ReferenceCopyLocalPaths in packageresolve.targets
makes us more consistent with the shipping targets but also starts
copying runtime assets for all dlls.  Add the rest of the shipping targets
that conditions this for dlls.

I've tried to make our logic as consistent with the shipping targets as I
can for now.