Show starting/finishing test progress in xunit console
When debugging problems in tests, it's very useful to be able to know which tests have started but haven't finished.  This commit adds a "-showprogress" option to the xunit.console.netcore.exe command line which, when set, will result in xunit tracing out to the console the starting and finishing of each individual test, e.g.

```
   System.Threading.Tasks.Dataflow.Tests.ActionBlockTests.TestPost [STARTING]
   System.Threading.Tasks.Dataflow.Tests.ActionBlockTests.TestPost [FINISHED] Time: 0.0316922s
```