Update BuildTools package layout and dependencies
Remove all the dependencies from the BuildTools package as they
don't make much sense. They will also hinder the move to installing
this package with the ExcludeVersion option on nuget as it will
install all the dependent packages as version-less as well which is
not what we want for our general library packages.

Rename Targets directory to PackageFiles because it contains more
things that just targets and also now always harvest that folder
recursively into the package lib folder now.