Support automatic nuget package creation for test packages
Support automatically creating nuget packages for test packages,
with the GenerateTestPackage msbuild property.  This brings in and
uses NuProj (https://github.com/nuproj/nuproj/).  The static nuspec
fields are stored in testpackage.nuproj, and the dynamic fields are
calculated and set by publishtest.targets.  Currently, the content of
the test package is the same as the test runtime directory without
dependencies.