Remove the workarounds in partialfacades.targets
These workarounds were in place because, in order to reference mscorlib in
our partial facades, we were referencing a Microsoft.DotNet.CoreCLR NuGet
package, which had a variety of extra assemblies that got automatically
pulled in as references. Also, the mscorlib we were referencing was
architecture-specific, which brought in a compiler warning that we
suppressed. None of this is necessary any longer now that we have
published proper reference assembly packages for mscorlib and the rest.