Add support for filtering resolved packages from targeting packages
This add support for filtering out resolved assets from the
targeting pack packages. It does this based on the following property
TargetingPackNugetPackageId which is set to the id of the package
to filter. It will remove any packages references from that package
except for the ones explicitly listed in the TargetingPackageReference
item group.