Add AssemblyResolver to load assemblies that would normally require BindingRedirects
MSBuild does not provide a convention for task dll's to specify bindingRedirects.
Sometimes a task may need to use two different libraries that use use different versions
of a common library (EG: Newtonsoft.JSON).  NuGet ensures that the higher version, is
used by the task, but one dependency will still have static reference to the lower version.

Without binding redirects this will result in a failure to load the assembly with the
lower version.  Address this by hooking AssemblyResolve and probing for the
assembly by simple name and choosing the first assembly that satisfies the request,
permitting higher versions.