Add initial support for reporting build statistic.
Generate a JSON file containing test build statistics and upload it to
DocumentDB for later processing. At present, the only statistic captured
is the number of successfully built tests.

Stop adding xunit and corerun zip packages to the list of downloads as
they are no longer required (they are pulled from packages.zip).