Eliminate nuproj dependency and refactor target sequence
This eliminates the dependency on nuproj by copying the targets
we were still using.  There is further trimming that can be done by
removing work that is unnecessary but I will do that later.

In addition to bringing these targets in I refactored the sequence
of the targets so that we can build project references without actually
having the binaries for the referenced project.  This will enable us to
better separate the x-plat builds so that packages can come from
from builds on multiple passes.