Consolidate copying of supplemental test data.
Add a new target, CopySupplementalTestData, that projects can use to copy
test data to the test directory. In addition, the task will also copy the
assets to the output directory so it can be picked up by the cloud test
infrastructure (it will be included in the test archive).
Exclude the test data from packages.zip as it's now being distributed with
the test binary.