Enable package build without binaries
In the closed package build we could assume that all platform binaries
would already be built, but in the open we strive to keep the platform
build minimal, EG: if I am building on OSX I don't need build Windows
libraries (at least not every build).  Despite not building every binary
we still want to build and validate the packages which have OSX
related assets, as well as the overall package authoring for the reference
packages to ensure package support for all applicable platforms.

To do this I have eliminated nearly all dependencies on reading
the built binaries in packaging.  The only remaining place where
we depend on the actual built binary is harvesting the dependencies
of the binary to write package dependencies.
I moved Generation validation (based on the dependencies) into
the library project build.  In doing this I eliminated the
dependency probing when walking the closure of references.
Since we are in the library project we get that full closure from NuGet.
To keep the full closure around I pulled OmitTransitiveCompileReferences
out of the task and performed it in MSBuild so that we could preserve
the transitive reference list.  This was something we needed to
do anyway since the public task doesn't have that functionality.