Add additional clean targets
Implementing these operations as MSBuild targets in BuildTools allows
them to be shared between clean.cmd and clean.sh, and used for both
CoreFX and CoreCLR. These targets are in addition to the Clean target
which is defined in build.proj in CoreFX and removes the bin directory.

Note that because MSBuild will soon be in the Tools directory, we
shouldn't have a build target that removes that directory. The logic to
remove that directory is (or will be) in the clean.cmd and clean.sh
scripts.