Split references by FrameworkAssembly/Dependency
Previously we'd try to determine if an assembly was mscorlib-based
or System.Runtime-based.  If mscorlib-based we assumed all
dependencies were framework assemblies.  This was incorrect, partial
facades may still reference packages and those packages may have an
OOB implementation.  This was recently the case for
System.IO.Compression's reference to System.Buffers.
https://github.com/dotnet/corefx/issues/5941