Add a simple preprocessor task
Add a very simple task that understands the limited form of #if/#endif
used by mscorlib.txt used for mscorlib's resources. This will allow us
to preprocess the file to generate a new .txt file that has only the
resources mscorlib needs (based on what FEATURE_ flags are defined)
and then use that as the .txt file MSBuild will use to generate
resources.

This will be used to resolve dotnet/coreclr#3090.