Specify specific version of S.R.Extensions
In CI, the OS X builds are failing intermittently because restoring the tool
runtime is pulling down an old version of the runtime assemblies for
System.Runtime.Extensions. This version expects some exports from
System.Native which have been renamed.

Setting the correct version of System.Runtime.Extensions seems to
address this issue (I tried restoring the tool-runtime from this
project.json a bunch of times and each time the right thing happened,
whereas without this fix it only worked about a quarter of the time).

Obviously there is some larger issue at play here as I would not expect
this to be needed, but I would really like to unblock our CI runs so
here's a pragmatic change which should do that.