Add a GenFacades build task
This change moves the core logic of GenFacades into a standalone library
called "GenFacades.Core". On top of that are two interfaces, the standard
GenFacades.exe, accepting the same arguments as before, and a new MSBuild
task which accepts similar parameters and which can be used directly from
MSBuild without the need to launch an entirely new process. This is
significantly faster and also avoids the need to re-interpret the command
line arguments in the console program, which are already known in the
MSBuild target which calls GenFacades.