Better error handling in GenFacades
I've removed the top-level catch block in the core GenFacades logic. That
error-handling logic has been moved to thhe respective wrapping projects
(GenFacades.exe and the MSBuild task assembly), which have a top-level
exception handler which logs the exception and returns an error code (or
returns false in the MSBuild task).

I've also added an MSBuild-Log trace listener to the GenFacades MSBuild
task. The core logic in GenFacades uses a lot of Trace messages to note
the status of the facade generation, like when a type is missing, or when
a duplicate type is detected. These messages are now visible in the
MSBuild log if any errors are encountered.

There was also a case in GenFacades where, in error conditions, we were
returning 'null' for the finished assembly, with the assumption that we
would error out somewhere higher up. Instead, I'm just explicitly throwing
an exception there.