Instead of caching the disabled analyzers list read it fresh for every analyzer.
What ends up happening before this change is that, when the vbcscompiler.exe server
is used, we end up sharing state between multiple instances of the analyzers.

And the state we end up sharing is the one that describes which analyzers should
be disabled.

The first time an analyzer runs, it will read the config file that describes which
analyzers should be disabled. And that value is cached. When the next analyzer
initializes, it will use the cached value even though the information in the file
might be different.

The fix is to not be too clever about reading this file. It should only contain a
handfull of analyzers so the cost of processing it should be low. We can revisit
this later if this becomes a perf bottleneck.